fix compilation problems
