+++ cmd/pow/cmd/storageinfo/get/get.go
@@ -1,4 +1,4 @@
-package cmd
+package get
 
 import (
 	"context"
@@ -6,31 +6,29 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	c "github.com/textileio/powergate/cmd/pow/common"
 	"google.golang.org/protobuf/encoding/protojson"
 )
 
-func init() {
-	storageInfoCmd.AddCommand(storageInfoGetCmd)
-}
-
-var storageInfoGetCmd = &cobra.Command{
+// Cmd is the command.
+var Cmd = &cobra.Command{
 	Use:   "get [cid]",
 	Short: "Returns the information about a stored cid.",
 	Long:  `Returns the information about a stored cid.`,
 	Args:  cobra.ExactArgs(1),
 	PreRun: func(cmd *cobra.Command, args []string) {
 		err := viper.BindPFlags(cmd.Flags())
-		checkErr(err)
+		c.CheckErr(err)
 	},
 	Run: func(cmd *cobra.Command, args []string) {
-		ctx, cancel := context.WithTimeout(context.Background(), cmdTimeout)
+		ctx, cancel := context.WithTimeout(context.Background(), c.CmdTimeout)
 		defer cancel()
 
-		res, err := powClient.StorageInfo.StorageInfo(mustAuthCtx(ctx), args[0])
-		checkErr(err)
+		res, err := c.PowClient.StorageInfo.StorageInfo(c.MustAuthCtx(ctx), args[0])
+		c.CheckErr(err)
 
 		json, err := protojson.MarshalOptions{Multiline: true, Indent: "  ", EmitUnpopulated: true}.Marshal(res.StorageInfo)
-		checkErr(err)
+		c.CheckErr(err)
 
 		fmt.Println(string(json))
 	},
