+++ cmd/pow/cmd/storageinfo/list/list.go
@@ -1,4 +1,4 @@
-package cmd
+package list
 
 import (
 	"context"
@@ -7,24 +7,22 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+	c "github.com/textileio/powergate/cmd/pow/common"
 	"google.golang.org/protobuf/encoding/protojson"
 )
 
-func init() {
-	storageInfoCmd.AddCommand(storageInfoListCmd)
-}
-
-var storageInfoListCmd = &cobra.Command{
+// Cmd is the command.
+var Cmd = &cobra.Command{
 	Use:   "list [optional cid1,cid2,...]",
 	Short: "Returns a list of information about all stored cids, filtered by cids if provided.",
 	Long:  `Returns a list of information about all stored cids, filtered by cids if provided.`,
 	Args:  cobra.MaximumNArgs(1),
 	PreRun: func(cmd *cobra.Command, args []string) {
 		err := viper.BindPFlags(cmd.Flags())
-		checkErr(err)
+		c.CheckErr(err)
 	},
 	Run: func(cmd *cobra.Command, args []string) {
-		ctx, cancel := context.WithTimeout(context.Background(), cmdTimeout)
+		ctx, cancel := context.WithTimeout(context.Background(), c.CmdTimeout)
 		defer cancel()
 
 		var cids []string
@@ -32,11 +30,11 @@ var storageInfoListCmd = &cobra.Command{
 			cids = strings.Split(args[0], ",")
 		}
 
-		res, err := powClient.StorageInfo.ListStorageInfo(mustAuthCtx(ctx), cids...)
-		checkErr(err)
+		res, err := c.PowClient.StorageInfo.ListStorageInfo(c.MustAuthCtx(ctx), cids...)
+		c.CheckErr(err)
 
 		json, err := protojson.MarshalOptions{Multiline: true, Indent: "  ", EmitUnpopulated: true}.Marshal(res)
-		checkErr(err)
+		c.CheckErr(err)
 
 		fmt.Println(string(json))
 	},
